Refactor into swc core and napi
