+++ packages/transformers/js/core/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+authors = ["Devon Govett <devongovett@gmail.com>"]
+name = "parcel-js-swc-core"
+version = "0.1.0"
+edition = "2018"
+
+[lib]
+crate-type = ["rlib"]
+
+[dependencies]
+swc_ecmascript = { version = "0.36.0", features = ["parser", "transforms", "module", "optimization", "react", "typescript", "utils", "visit", "codegen", "utils"] }
+swc_ecma_preset_env = "0.20.1"
+swc_common = { version = "0.10.19", features = ["tty-emitter", "sourcemap"] }
+swc_atoms = "0.2.6"
+indoc = "1.0.3"
+serde = "1.0.123"
+Inflector = "0.11.4"
+data-encoding = "2.3.2"
+sha-1 = "0.9.4"
+dunce = "1.0.1"
