+++ packages/transformers/js/core/src/decl_collector.rs
