+++ packages/transformers/js/core/src/modules.rs
