+++ packages/transformers/js/core/src/utils.rs
