+++ packages/transformers/js/napi/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
 authors = ["Devon Govett <devongovett@gmail.com>"]
-name = "parcel-js-swc"
+name = "parcel-js-swc-napi"
 version = "0.1.0"
+edition = "2018"
 
 [lib]
 crate-type = ["cdylib"]
@@ -9,16 +10,7 @@ crate-type = ["cdylib"]
 [dependencies]
 napi = { version = "1", features = ["serde-json"] }
 napi-derive = "1"
-swc_ecmascript = { version = "0.36.0", features = ["parser", "transforms", "module", "optimization", "react", "typescript", "utils", "visit", "codegen", "utils"] }
-swc_ecma_preset_env = "0.20.1"
-swc_common = { version = "0.10.19", features = ["tty-emitter", "sourcemap"] }
-swc_atoms = "0.2.6"
-indoc = "1.0.3"
-serde = "1.0.123"
-Inflector = "0.11.4"
-data-encoding = "2.3.2"
-sha-1 = "0.9.4"
-dunce = "1.0.1"
+parcel-js-swc-core = { path = "../core" }
 
 [target.'cfg(target_os = "macos")'.dependencies]
 jemallocator = { version = "0.3.2", features = ["disable_initial_exec_tls"] }
